requirements.txt
# Core AI Framework
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.20
openai==1.12.0

# Composio Tool Router
composio-core==0.5.8
composio-langchain==0.5.8

# Voice Processing (we'll add later if needed)
# openai-whisper==20231117
# elevenlabs==0.2.27

# Code Analysis
ast-comments==1.1.2

# Web & API
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
aiohttp==3.9.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging
loguru==0.7.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.0

# .\venv\Scripts\Activate.ps1